Move buffer allocation abstraction down to plain malloc()/free().
