+++ src/app/plugin/test-on-off/Makefile
@@ -1,6 +1,14 @@
 NAME=ChipZclOnOffTestDriver
 GEN=../../gen
-REQUIRED_PLUGINS=../core-data-model/zcl-data-model.o ../cluster-server-on-off/on-off-server.o $(GEN)/gen-callback-stubs.o $(GEN)/gen-specs.o $(GEN)/gen-command-handler.o ../binding-mock/mock.o cluster-cmd-on-off-test.o
+REQUIRED_PLUGINS=\
+  ../core-data-model/zcl-data-model.o \
+  ../core-api/core-api.o \
+  ../cluster-server-on-off/on-off-server.o \
+  ../binding-mock/mock.o \
+  cluster-cmd-on-off-test.o \
+  $(GEN)/gen-callback-stubs.o \
+  $(GEN)/gen-specs.o \
+  $(GEN)/gen-command-handler.o
 
 # ---------- GENERIC -------------
 CC=gcc
